+++ pymc4/flow/executor.py
@@ -1,17 +1,13 @@
 import types
-import functools
-from typing import Any, Tuple, Dict, Union, List, Optional, Set, Mapping
-from collections import ChainMap
-import itertools
+from typing import Any, Tuple, Dict, Union, Optional, Mapping
 
 import tensorflow as tf
-from tensorflow_probability.python.internal import prefer_static
-
 import pymc4 as pm
 from pymc4 import coroutine_model
 from pymc4 import scopes
-from pymc4 import utils
 from pymc4.distributions import distribution
+
+from pymc4.flow import SamplingState
 
 
 ModelType = Union[types.GeneratorType, coroutine_model.Model]
@@ -22,20 +18,6 @@ MODEL_POTENTIAL_AND_DETERMINISTIC_TYPES = (
     distribution.Potential,
     distribution.Deterministic,
 )
-
-
-def _chain_map_iter(self):
-    """Keep ordering of maps on Python3.6.
-    See https://bugs.python.org/issue32792
-    Once Python3.6 is not supported, this can be deleted.
-    """
-    d = {}
-    for mapping in reversed(self.maps):
-        d.update(mapping)  # reuses stored hash values if possible
-    return iter(d)
-
-
-ChainMap.__iter__ = _chain_map_iter  # type: ignore
 
 
 class EvaluationError(RuntimeError):
@@ -98,354 +80,6 @@ class StopExecution(StopIteration):
 
 class EarlyReturn(StopIteration):
     ...
-
-
-class SamplingState:
-    __slots__ = (
-        "transformed_values",
-        "untransformed_values",
-        "transformed_values_batched",
-        "untransformed_values_batched",
-        "observed_values",
-        "posterior_predictives",
-        "all_values",
-        "all_unobserved_values",
-        "all_unobserved_values_batched",
-        "distributions",
-        "prior_distributions",
-        "likelihood_distributions",
-        "potentials",
-        "deterministics",
-    )
-
-    def __init__(
-        self,
-        transformed_values: Dict[str, Any] = None,
-        untransformed_values: Dict[str, Any] = None,
-        observed_values: Dict[str, Any] = None,
-        distributions: Dict[str, distribution.Distribution] = None,
-        potentials: List[distribution.Potential] = None,
-        deterministics: Dict[str, Any] = None,
-        posterior_predictives: Optional[Set[str]] = None,
-        transformed_values_batched: Dict[str, Any] = None,
-        untransformed_values_batched: Dict[str, Any] = None,
-    ) -> None:
-        # verbose __init__
-        if transformed_values is None:
-            transformed_values = dict()
-        else:
-            transformed_values = transformed_values.copy()
-        if untransformed_values is None:
-            untransformed_values = dict()
-        else:
-            untransformed_values = untransformed_values.copy()
-        if transformed_values_batched is None:
-            transformed_values_batched = dict()
-        else:
-            transformed_values_batched = transformed_values_batched.copy()
-        if untransformed_values_batched is None:
-            untransformed_values_batched = dict()
-        else:
-            untransformed_values_batched = untransformed_values_batched.copy()
-        if observed_values is None:
-            observed_values = dict()
-        else:
-            observed_values = observed_values.copy()
-        if distributions is None:
-            distributions = dict()
-        else:
-            distributions = distributions.copy()
-        if potentials is None:
-            potentials = list()
-        else:
-            potentials = potentials.copy()
-        if deterministics is None:
-            deterministics = dict()
-        else:
-            deterministics = deterministics.copy()
-        if posterior_predictives is None:
-            posterior_predictives = set()
-        else:
-            posterior_predictives = posterior_predictives.copy()
-        self.transformed_values = transformed_values
-        self.untransformed_values = untransformed_values
-        self.observed_values = observed_values
-
-        self.transformed_values_batched = transformed_values_batched
-        self.untransformed_values_batched = untransformed_values_batched
-
-        self.all_values = ChainMap(
-            self.untransformed_values, self.transformed_values, self.observed_values
-        )
-
-        self.all_unobserved_values = ChainMap(self.untransformed_values, self.transformed_values)
-        self.all_unobserved_values_batched = ChainMap(
-            self.transformed_values_batched, self.untransformed_values_batched
-        )
-
-        self.distributions = distributions
-        self.prior_distributions = {}
-        self.likelihood_distributions = {}
-
-        self.potentials = potentials
-        self.deterministics = deterministics
-        self.posterior_predictives = posterior_predictives
-
-    def collect_log_prob_elemwise(self):
-        return itertools.chain(
-            (dist.log_prob(self.all_values[name]) for name, dist in self.distributions.items()),
-            (p.value for p in self.potentials),
-        )
-
-    def collect_log_prob(self, is_reduced=True):
-        if is_reduced:
-            return sum(map(prefer_static.reduce_sum, self.collect_log_prob_elemwise()))
-        else:
-            return sum(self.collect_log_prob_elemwise())
-
-    def _collect_log_prob_elemwise_lp_prior(self, distrs):
-        """
-            Collects log probabilities for prior variables in sMC
-        """
-
-        def reduce_sum_smc(value, ind_exclude):
-            # `ind_exluce` argument determines the index that
-            # should be kept when reducing `value` tensor.
-            ndim = len(value.shape)
-            return prefer_static.reduce_sum(
-                value, axis=[*range(0, ind_exclude), *range(ind_exclude + 1, ndim)]
-            )
-
-        for name, dist in distrs.items():
-            # TODO: for now if we are setting explicit batch dim in sMC and
-            # distribution has plate parameter (batch_stack in pymc4) then the distribution
-            # expects the sample of shape: [*batch_stack, draws, *event_shape] for `log_prob`
-            # but the values are stored as [draws,...] because of the sequential_monte_carlo
-            # op design in tfp. For now we can only check for batch_stack in distribution,
-            # transpose the value and set correct axises to apply `reduce_sum`
-            # This potentially can be solved with the additional argument passed to the
-            # distribution initialization alongside the `batch_stack` for the BatchStacker.
-            value = self.all_values[name]
-            val_ndim = len(value.shape)
-            int_exclude = 0
-            if dist.batch_stack:
-                bs = 1 if isinstance(dist.batch_stack, int) else len(dist.batch_stack)
-                value = tf.transpose(value, [*range(1, bs + 1), 0, *range(bs + 1, val_ndim)])
-                int_exclude = bs
-            yield reduce_sum_smc(dist.log_prob(value), ind_exclude=int_exclude)
-
-    def _collect_log_prob_elemwise_lp_lkh(self, distrs):
-        for name, dist in distrs.items():
-            # For sMC we need to move first axis (draws axis) to the right position
-            # so that the shapes of distribution parameters and value can be broadcasted
-            value_st = self.all_values[name]
-            val_dims, val_ndim = value_st.shape, value_st.ndim
-            dist_ndim = dist.batch_shape.rank + dist.event_shape.rank
-            ndim_diff = prefer_static.abs(val_ndim - dist_ndim)
-
-            # if...else supported by tf.function
-            if val_ndim > dist_ndim:
-                # If value ndim is larger thatn ndim of distribution params then
-                # we are transposing the value tensor
-                min_diff_ndim = prefer_static.minimum(ndim_diff + 1, val_ndim)
-                axis_transpose = prefer_static.concat(
-                    [
-                        prefer_static.range(1, min_diff_ndim),
-                        tf.constant([0]),
-                        prefer_static.range(min_diff_ndim, val_ndim),
-                    ],
-                    axis=0,
-                )
-                value = tf.transpose(value_st, axis_transpose)
-                reduce_axis = prefer_static.concat(
-                    [
-                        prefer_static.range(0, min_diff_ndim - 1),
-                        prefer_static.range(min_diff_ndim, val_ndim),
-                    ],
-                    axis=0,
-                )
-            elif val_ndim < dist_ndim:
-                # Otherwise we are reshaping the value tensor to make it broadcastable
-                shape_ = prefer_static.concat(
-                    [
-                        tf.constant([val_dims[0]]),
-                        prefer_static.ones(ndim_diff, dtype=tf.int32),  # indices are int32
-                        tf.constant([*val_dims[1:]], dtype=tf.int32),
-                    ],
-                    axis=0,
-                )
-                value = prefer_static.reshape(value_st, shape=shape_)
-                reduce_axis = prefer_static.range(1, dist_ndim)
-            else:
-                reduce_axis = prefer_static.range(1, val_ndim)
-                value = value_st
-
-            yield prefer_static.reduce_sum(
-                dist.log_prob(value), axis=reduce_axis,
-            )
-
-    def collect_log_prob_smc(self, is_prior):
-        """
-            Collects log probabilities for likelihood variables in sMC.
-            Since sMC requires the `draws` dimension to be kept explicitly
-            while the graph is evaluated, we can't combine sMC prbability
-            collection with the NUTS log probability collection.
-        """
-        # Because of the different logic with reduce sum
-        # it is not possible to merge two log prob ops
-        # so for now we are separating them with the logic
-        # of executing smc log ops and sampling log prob ops
-        if is_prior is True:
-            distrs = self.prior_distributions
-            log_prob_elemwise = self._collect_log_prob_elemwise_lp_prior(distrs)
-        else:
-            distrs = self.likelihood_distributions
-            log_prob_elemwise = self._collect_log_prob_elemwise_lp_lkh(distrs)
-            log_prob_elemwise = itertools.chain(
-                log_prob_elemwise, (p.value for p in self.potentials)
-            )
-        log_prob = sum(log_prob_elemwise)
-        return log_prob
-
-    def __repr__(self):
-        def get_distribution_class_names(distribution_dict):
-            return ["{}:{}".format(d.__class__.__name__, k) for k, d in distribution_dict.items()]
-
-        # display keys only
-        untransformed_values = list(self.untransformed_values)
-        transformed_values = list(self.transformed_values)
-        observed_values = list(self.observed_values)
-        deterministics = list(self.deterministics)
-        posterior_predictives = list(self.posterior_predictives)
-        # format like dist:name prior_distribution, likelihood_distribution
-        prior_distributions = get_distribution_class_names(self.prior_distributions)
-        likelihood_distributions = get_distribution_class_names(self.likelihood_distributions)
-        # be less verbose here
-        num_potentials = len(self.potentials)
-
-        indent = 4 * " "
-        return (
-            "{}(\n"
-            + indent
-            + "untransformed_values: {}\n"
-            + indent
-            + "transformed_values: {}\n"
-            + indent
-            + "observed_values: {}\n"
-            + indent
-            + "prior_distributions: {}\n"
-            + indent
-            + "likelihood_distributions: {}\n"
-            + indent
-            + "num_potentials={}\n"
-            + indent
-            + "deterministics: {}\n"
-            + indent
-            + "posterior_predictives: {})"
-        ).format(
-            self.__class__.__name__,
-            untransformed_values,
-            transformed_values,
-            observed_values,
-            prior_distributions,
-            likelihood_distributions,
-            num_potentials,
-            deterministics,
-            posterior_predictives,
-        )
-
-    @classmethod
-    def from_values(
-        cls, values: Dict[str, Any] = None, observed_values: Dict[str, Any] = None
-    ) -> "SamplingState":
-        if values is None:
-            return cls(observed_values=observed_values)
-        transformed_values = dict()
-        untransformed_values = dict()
-        transformed_values_batched = dict()
-        untransformed_values_batched = dict()
-        # split by `nest/name` or `nest/__transform_name`
-        for fullname in values:
-            namespec = utils.NameParts.from_name(fullname)
-            if namespec.is_transformed:
-                transformed_values[fullname] = values[fullname]
-            else:
-                untransformed_values[fullname] = values[fullname]
-        return cls(
-            transformed_values,
-            untransformed_values,
-            observed_values,
-            transformed_values_batched=transformed_values_batched,
-            untransformed_values_batched=untransformed_values_batched,
-        )
-
-    def clone(self) -> "SamplingState":
-        return self.__class__(
-            transformed_values=self.transformed_values,
-            untransformed_values=self.untransformed_values,
-            observed_values=self.observed_values,
-            distributions=self.distributions,
-            potentials=self.potentials,
-            deterministics=self.deterministics,
-            posterior_predictives=self.posterior_predictives,
-        )
-
-    def as_sampling_state(self) -> "Tuple[SamplingState, List[str]]":
-        """Create a sampling state that should be used within MCMC sampling.
-        There are some principles that hold for the state.
-            1. Check there is at least one distribution
-            2. Check all transformed distributions are autotransformed
-            3. Remove untransformed values if transformed are present
-            4. Remove all other irrelevant values
-        """
-        if not self.distributions:
-            raise TypeError(
-                "No distributions found in the state. "
-                "the model you evaluated is empty and does not yield any PyMC4 distribution"
-            )
-        untransformed_values = dict()
-        transformed_values = dict()
-        need_to_transform_after = list()
-        observed_values = dict()
-
-        for name, dist in self.distributions.items():
-            namespec = utils.NameParts.from_name(name)
-            if dist.transform is not None and name not in self.observed_values:
-                transformed_namespec = namespec.replace_transform(dist.transform.name)
-                if transformed_namespec.full_original_name not in self.transformed_values:
-                    raise TypeError(
-                        "Transformed value {!r} is not found for {} distribution with name {!r}. "
-                        "You should evaluate the model using the transformed executor to get "
-                        "the correct sampling state.".format(
-                            transformed_namespec.full_original_name, dist, name
-                        )
-                    )
-                else:
-                    transformed_values[
-                        transformed_namespec.full_original_name
-                    ] = self.transformed_values[transformed_namespec.full_original_name]
-                    need_to_transform_after.append(transformed_namespec.full_untransformed_name)
-            else:
-                if name in self.observed_values:
-                    observed_values[name] = self.observed_values[name]
-                elif name in self.untransformed_values:
-                    untransformed_values[name] = self.untransformed_values[name]
-                else:
-                    raise TypeError(
-                        "{} distribution with name {!r} does not have the corresponding value "
-                        "in the state. This may happen if the current "
-                        "state was modified in the wrong way."
-                    )
-        return (
-            self.__class__(
-                transformed_values=transformed_values,
-                untransformed_values=untransformed_values,
-                transformed_values_batched=self.transformed_values_batched.copy(),
-                untransformed_values_batched=self.untransformed_values_batched.copy(),
-                observed_values=observed_values,
-            ),
-            need_to_transform_after,
-        )
 
 
 # when we make changes in a subclass we usually call super() that will require self to be present in signature
